Disney BRDF does not need to be visible.
